# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hydrogen Laravel Generator APIs",
        "contact": {
            "email": "amrmohamed257@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/carts?include=user&cartItems": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "getCartList",
                "description": "Get all Cart",
                "operationId": "0aa1ca6d616a34dc2baae785218e01f5",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "carts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cart"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/carts": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "createCart",
                "description": "Create Cart",
                "operationId": "12ef9e838deaacc8deb640aa4c08c508",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cart"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cart": {
                                            "$ref": "#/components/schemas/Cart"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/carts/{id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "getCartItem",
                "description": "Get Cart",
                "operationId": "7bd9e3d21a0d450bf1ad7940e7f909cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Cart",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cart": {
                                            "$ref": "#/components/schemas/Cart"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "updateCart",
                "description": "Update Cart",
                "operationId": "59e1c6fe19a657d8bfcd2d7eddd4d214",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Cart",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cart"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "cart": {
                                            "$ref": "#/components/schemas/Cart"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "deleteCart",
                "description": "Delete Cart",
                "operationId": "fdeff13c06a1dc9b90c25dc79bc5fe4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Cart",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cart": {
                "required": [],
                "properties": {
                    "user_id": {
                        "nullable": false,
                        "description": "Owner id of the cart",
                        "type": "integer",
                        "readOnly": false
                    },
                    "created_at": {
                        "nullable": true,
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "deleted_at": {
                        "nullable": true,
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "nullable": true,
                        "schema": "Cart",
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}